requirements.txt
streamlit==1.39.0
google-generativeai==0.6.0
PyPDF2==3.0.1
Pillow==10.4.0
gTTS==2.5.1
requests==2.32.3
